+++ Pandora/Paradigm/Inventory/Optics.hs
@@ -8,10 +8,8 @@ import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Representable (Representable (Representation, (<#>), tabulate))
 import Pandora.Pattern.Object.Setoid (Setoid ((==)))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run)
-import Pandora.Paradigm.Primary.Functor.Function ((!))
 import Pandora.Paradigm.Primary.Functor.Identity (Identity (Identity))
-import Pandora.Paradigm.Primary.Functor.Maybe (Maybe (Just, Nothing))
-import Pandora.Paradigm.Primary.Functor.Product (Product ((:*:)), attached)
+import Pandora.Paradigm.Primary.Functor.Product (Product ((:*:)))
 import Pandora.Paradigm.Primary.Object.Boolean ((?))
 import Pandora.Paradigm.Inventory.Store (Store (Store), position, look, retrofit)
 import Pandora.Paradigm.Schemes.PQ_ (PQ_ (PQ_))
@@ -20,10 +18,12 @@ import Pandora.Paradigm.Schemes.P_T (P_T (P_T))
 infixr 0 :-.
 infixr 0 :~.
 
+type Optics mode = PQ_ (->) (P_T Store mode)
+
 type (:-.) src tgt = Lens src tgt
 
 -- Reference to taret within some source
-type Lens = PQ_ (->) (P_T Store Identity)
+type Lens = Optics Identity
 
 instance Category Lens where
 	identity = PQ_ $ \src -> P_T . Store $ Identity src :*: identity . extract
@@ -47,17 +47,3 @@ over lens f = extract . retrofit (f <$>) . run . run lens
 -- | Representable based lens
 represent :: (Representable t, Setoid (Representation t)) => Representation t -> t a :-. a
 represent r = PQ_ $ \x -> P_T $ Store $ Identity (r <#> x) :*: \new -> tabulate (\r' -> r' == r ? extract new $ r' <#> x)
-
-type Prism = PQ_ (->) (P_T Store Maybe)
-
-instance Category Prism where
-	identity = PQ_ $ \src -> P_T . Store $ Just src :*: (src !)
-	PQ_ to . PQ_ from = PQ_ $ \src -> P_T $
-		case attached (run (run (from src))) of
-			Just focused -> src <$ run (to focused)
-			Nothing -> Store $ Nothing :*: (src !)
-
-preview :: Prism src tgt -> src -> Maybe tgt
-preview prism = position . run . run prism
-
-type Optics mod = PQ_ (->) (P_T Store mod)